+++ src/Data/Chimera/ContinuousMapping.hs
@@ -1,3 +1,8 @@
+{-# LANGUAGE CPP #-}
+{-# LANGUAGE DerivingStrategies #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
+{-# LANGUAGE TypeApplications #-}
+
 -- |
 -- Module:      Data.Chimera.ContinuousMapping
 -- Copyright:   (c) 2017 Andrew Lelechenko
@@ -63,25 +68,36 @@
 -- Namely, cast the state to 'Word' @->@ 'Bool', ready for memoization:
 --
 -- > cast :: (Int -> Int -> Bool) -> (Word -> Bool)
--- > cast f = \n -> let (x, y) = fromZCurve n in
--- >  f (wordToInt x) (wordToInt y)
+-- > cast f = \n -> let (x, y) = fromZCurve n in f (fromHalf x) (fromHalf y)
+-- >   where
+-- >     fromHalf :: HalfWord -> Int
+-- >     fromHalf = wordToInt . fromIntegral @HalfWord @Word
 --
 -- and then back:
 --
 -- > uncast :: (Word -> Bool) -> (Int -> Int -> Bool)
--- > uncast g = \x y -> g (toZCurve (intToWord x) (intToWord y))
+-- > uncast g = \x y -> g (toZCurve (toHalf x) (toHalf y))
+-- >   where
+-- >     toHalf :: Int -> HalfWord
+-- >     toHalf = fromIntegral @Word @HalfWord . intToWord
 module Data.Chimera.ContinuousMapping (
   intToWord,
   wordToInt,
+  HalfWord,
   toZCurve,
   fromZCurve,
+  ThirdWord,
   toZCurve3,
   fromZCurve3,
 ) where
 
+import Data.Bifunctor
 import Data.Bits
 import Data.Chimera.FromIntegral
+import Data.Coerce
 import Data.Word
+
+#include "MachDeps.h"
 
 -- | Total map, which satisfies
 --
@@ -111,16 +127,27 @@ wordToInt :: Word -> Int
 wordToInt w = word2int $ (if w .&. 1 == 0 then id else complement) (w `shiftR` 1)
 {-# INLINE wordToInt #-}
 
+-- | 32 bits on 64-bit architecture, 16 bits on 32-bit architecture.
+--
+-- To create a value of type 'HalfWord' use 'fromIntegral'.
+--
+-- @since 0.4.0.0
+#if WORD_SIZE_IN_BITS == 64
+newtype HalfWord = HalfWord Word32
+  deriving newtype (Eq, Ord, Show, Read, Bits, FiniteBits, Bounded, Enum, Num, Integral, Real)
+#else
+newtype HalfWord = HalfWord Word16
+  deriving newtype (Eq, Ord, Show, Read, Bits, FiniteBits, Bounded, Enum, Num, Integral, Real)
+#endif
+
 -- | Total map from plain to line, continuous almost everywhere.
 -- See <https://en.wikipedia.org/wiki/Z-order_curve Z-order curve>.
---
--- Only lower halfs of bits of arguments are used (32 bits on 64-bit architecture).
 --
 -- >>> [ toZCurve x y | x <- [0..3], y <- [0..3] ]
 -- [0,2,8,10,1,3,9,11,4,6,12,14,5,7,13,15]
 --
 -- @since 0.2.0.0
-toZCurve :: Word -> Word -> Word
+toZCurve :: HalfWord -> HalfWord -> Word
 toZCurve x y = part1by1 y `shiftL` 1 .|. part1by1 x
 
 -- | Inverse for 'toZCurve'.
@@ -130,20 +157,80 @@ toZCurve x y = part1by1 y `shiftL` 1 .|. part1by1 x
 -- [(0,0),(1,0),(0,1),(1,1),(2,0),(3,0),(2,1),(3,1),(0,2),(1,2),(0,3),(1,3),(2,2),(3,2),(2,3),(3,3)]
 --
 -- @since 0.2.0.0
-fromZCurve :: Word -> (Word, Word)
+fromZCurve :: Word -> (HalfWord, HalfWord)
 fromZCurve z = (compact1by1 z, compact1by1 (z `shiftR` 1))
+
+-- | 21 bits on 64-bit architecture, 10 bits on 32-bit architecture.
+--
+-- To create a value of type 'ThirdWord' use 'fromIntegral'.
+--
+-- @since 0.4.0.0
+newtype ThirdWord = ThirdWord Word32
+  deriving newtype (Eq, Ord, Show)
+
+mkThirdWord :: Word32 -> ThirdWord
+mkThirdWord n = t
+  where
+    t = ThirdWord (n .&. ((1 `shiftL` finiteBitSize t) - 1))
+
+instance Read ThirdWord where
+  readsPrec = (fmap (first mkThirdWord) .) . readsPrec
+
+instance Bits ThirdWord where
+  (.&.) = coerce ((.&.) @Word32)
+  (.|.) = coerce ((.|.) @Word32)
+  xor = coerce (xor @Word32)
+  complement (ThirdWord n) = mkThirdWord (complement n)
+  shift (ThirdWord n) k = mkThirdWord (shift n k)
+  bitSize = finiteBitSize
+  bitSizeMaybe = Just . finiteBitSize
+  isSigned = coerce (isSigned @Word32)
+  testBit = coerce (testBit @Word32)
+  bit = mkThirdWord . bit
+  popCount = coerce (popCount @Word32)
+
+  rotate t k'
+    | k == 0 = t
+    | otherwise = (t `shiftL` k) .|. (t `shiftR` (fbs - k))
+    where
+      fbs = finiteBitSize t
+      k = k' `mod` fbs
+
+instance FiniteBits ThirdWord where
+  finiteBitSize = const $ finiteBitSize (0 :: Word) `quot` 3
+
+instance Bounded ThirdWord where
+  minBound = mkThirdWord minBound
+  maxBound = mkThirdWord maxBound
+
+instance Enum ThirdWord where
+  toEnum = mkThirdWord . toEnum
+  fromEnum = coerce (fromEnum @Word32)
+
+instance Num ThirdWord where
+  ThirdWord x + ThirdWord y = mkThirdWord (x + y)
+  ThirdWord x * ThirdWord y = mkThirdWord (x * y)
+  negate (ThirdWord x) = mkThirdWord (negate x)
+  abs = coerce (abs @Word32)
+  signum = coerce (signum @Word32)
+  fromInteger = mkThirdWord . fromInteger
+
+instance Real ThirdWord where
+  toRational = coerce (toRational @Word32)
+
+instance Integral ThirdWord where
+  quotRem = coerce (quotRem @Word32)
+  toInteger = coerce (toInteger @Word32)
 
 -- | Total map from space to line, continuous almost everywhere.
 -- See <https://en.wikipedia.org/wiki/Z-order_curve Z-order curve>.
---
--- Only lower thirds of bits of arguments are used (21 bits on 64-bit architecture).
 --
 -- >>> [ toZCurve3 x y z | x <- [0..3], y <- [0..3], z <- [0..3] ]
 -- [0,4,32,36,2,6,34,38,16,20,48,52,18,22,50,54,1,5,33,37,3,7,35,39,17,21,49,53,19,23,51,55,
 -- 8,12,40,44,10,14,42,46,24,28,56,60,26,30,58,62,9,13,41,45,11,15,43,47,25,29,57,61,27,31,59,63]
 --
 -- @since 0.2.0.0
-toZCurve3 :: Word -> Word -> Word -> Word
+toZCurve3 :: ThirdWord -> ThirdWord -> ThirdWord -> Word
 toZCurve3 x y z = part1by2 z `shiftL` 2 .|. part1by2 y `shiftL` 1 .|. part1by2 x
 
 -- | Inverse for 'toZCurve3'.
@@ -156,11 +243,11 @@ toZCurve3 x y z = part1by2 z `shiftL` 2 .|. part1by2 y `shiftL` 1 .|. part1by2 x
 --  (0,2,2),(1,2,2),(0,3,2),(1,3,2),(0,2,3),(1,2,3),(0,3,3),(1,3,3),(2,2,2),(3,2,2),(2,3,2),(3,3,2),(2,2,3),(3,2,3),(2,3,3),(3,3,3)]
 --
 -- @since 0.2.0.0
-fromZCurve3 :: Word -> (Word, Word, Word)
+fromZCurve3 :: Word -> (ThirdWord, ThirdWord, ThirdWord)
 fromZCurve3 z = (compact1by2 z, compact1by2 (z `shiftR` 1), compact1by2 (z `shiftR` 2))
 
 -- Inspired by https://fgiesen.wordpress.com/2009/12/13/decoding-morton-codes/
-part1by1 :: Word -> Word
+part1by1 :: HalfWord -> Word
 part1by1 x = fromIntegral (x5 :: Word64)
   where
     x0 = fromIntegral x .&. 0x00000000ffffffff
@@ -171,7 +258,7 @@ part1by1 x = fromIntegral (x5 :: Word64)
     x5 = (x4 `xor` (x4 `shiftL` 1)) .&. 0x5555555555555555
 
 -- Inspired by https://fgiesen.wordpress.com/2009/12/13/decoding-morton-codes/
-part1by2 :: Word -> Word
+part1by2 :: ThirdWord -> Word
 part1by2 x = fromIntegral (x5 :: Word64)
   where
     x0 = fromIntegral x .&. 0x00000000ffffffff
@@ -182,7 +269,7 @@ part1by2 x = fromIntegral (x5 :: Word64)
     x5 = (x4 `xor` (x4 `shiftL` 2)) .&. 0x1249249249249249
 
 -- Inspired by https://fgiesen.wordpress.com/2009/12/13/decoding-morton-codes/
-compact1by1 :: Word -> Word
+compact1by1 :: Word -> HalfWord
 compact1by1 x = fromIntegral (x5 :: Word64)
   where
     x0 = fromIntegral x .&. 0x5555555555555555
@@ -193,7 +280,7 @@ compact1by1 x = fromIntegral (x5 :: Word64)
     x5 = (x4 `xor` (x4 `shiftR` 16)) .&. 0x00000000ffffffff
 
 -- Inspired by https://fgiesen.wordpress.com/2009/12/13/decoding-morton-codes/
-compact1by2 :: Word -> Word
+compact1by2 :: Word -> ThirdWord
 compact1by2 x = fromIntegral (x5 :: Word64)
   where
     x0 = fromIntegral x .&. 0x1249249249249249